% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totemData.R
\docType{class}
\name{totemData-class}
\alias{totemData-class}
\alias{totemDataFromDataFrame}
\title{totemData object and constructors}
\usage{
totemDataFromDataFrame(df)
}
\arguments{
\item{df}{a 'data.frame' object with the subjects in the rows and their attributes in the columns}
}
\value{
A totemData object
}
\description{
\code{totemData} is used to store the TOTEM representation
}
\section{Slots}{

\describe{
\item{\code{data}}{a 'tibble' representing entities, their counts (column 'count') and relative frequencies (column 'freq').}

\item{\code{iProjections}}{a 'list' of iProjections}
}}

