% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iProjection.R
\docType{class}
\name{iProjection-class}
\alias{iProjection-class}
\alias{iProjection}
\title{iProjection class}
\usage{
iProjection(totemData, design, reference = NULL)
}
\arguments{
\item{totemData}{totemData object with the empirical distribution.}

\item{design}{a \code{formula}, expresses how the target (left hand side) should depend
on attributes (right hand side).}

\item{reference}{reference distribution (default: uniform distribution)}
}
\description{
The \code{iProjection} class represents an \eqn{i}-projection from a reference distribution.
}
\section{Slots}{

\describe{
\item{\code{design}}{design formula describing the conditions}

\item{\code{conditions}}{values resulting from applying conditions onto the data}

\item{\code{mat}}{matrix representation of the conditions}

\item{\code{iProjection}}{i-projection}

\item{\code{reference}}{reference distribution}
}}

\author{
Orestis Loukas & Ho Ryun Chung
}
